change amount to bigDecimal

+++ src/main/java/model/AccountStatement.java
@@ -1,5 +1,6 @@
 package model;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 public class AccountStatement {
@@ -7,9 +8,9 @@ public class AccountStatement {
     private final double currentBalance;
     private final BankOperationType bankOperationType;
     private final ZonedDateTime date;
-    private final double amount;
+    private final BigDecimal amount;
 
-    public AccountStatement(BankOperationType bankOperationType, ZonedDateTime date, double amount, double currentBalance) {
+    public AccountStatement(BankOperationType bankOperationType, ZonedDateTime date, BigDecimal amount, double currentBalance) {
         this.bankOperationType = bankOperationType;
         this.date = date;
         this.amount = amount;
@@ -20,7 +21,7 @@ public class AccountStatement {
         return bankOperationType;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
